requirements.txt
asgiref==3.2.7
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-azure-sql-backend==2.1.1.1
django-cors-headers==3.3.0
django-pyodbc-azure==2.1.0.0
django-storages==1.9.1
djangorestframework==3.11.0
idna==2.9
msal==1.3.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
